+++ src/main/java/com/alvaro/empresas/passagens/autobuses/repositories/PisoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PisoRepository extends JpaRepository<PisoModel, Long> {
+public interface PisoRepository extends JpaRepository<PisoModel, Integer> {
 }
