+++ src/main/java/com/alvaro/empresas/passagens/autobuses/services/PosicionIndisponibleService.java
@@ -27,7 +27,7 @@ public class PosicionIndisponibleService {
     }
 
     public void deleteAll(List<PosicionIndisponibleModel> list) {
-        List<Long> ids = new ArrayList<>();
+        List<Integer> ids = new ArrayList<>();
         list.forEach(model -> {
             ids.add(model.getId());
         });
